+++ Sources/BFKit/Apple/UIKit/UIViewExtension.swift
@@ -170,7 +170,9 @@ public extension UIView {
     /// - Parameters:
     ///   - colors: Array of UIColor instances.
     ///   - direction: Direction of the gradient.
-    public func gradient(colors: [UIColor], direction: UIViewGradientDirection) {
+    /// - Returns: Returns the created CAGradientLayer.
+    @discardableResult
+    public func gradient(colors: [UIColor], direction: UIViewGradientDirection) -> CAGradientLayer {
         let gradient: CAGradientLayer = CAGradientLayer()
         gradient.frame = self.bounds
         
@@ -205,6 +207,8 @@ public extension UIView {
             gradient.endPoint = endPoint
         }
         self.layer.insertSublayer(gradient, at:0)
+        
+        return gradient
     }
     
     /// Create a smooth linear gradient, requires more computational time than 
@@ -476,6 +480,7 @@ extension UIView {
 // MARK: - UIView inspectable extension
 
 /// Extends UIView with inspectable variables.
+@IBDesignable
 extension UIView {
     // MARK: - Variables